Move lib package
